requirements.txt
langchain
openai
anthropic
python-telegram-bot
PyMuPDF
faiss-cpu
python-dotenv
langchain-openai
langchain-community
psycopg2